refactor(comprehension): extract llm tests to sibling file
Move the single inline `#[cfg(test)] mod tests` out of
primer-comprehension/src/llm.rs into a sibling llm/tests.rs
(non-root module -> one dir deeper, mkdir), following the
macos26/analyzer precedent (#295).

Pure verbatim test-module relocation (plus the rustfmt re-wraps the
one-level dedent triggers): 717 -> 317 + 399 lines. The test module
stays a child of `llm`, so `use super::*` still reaches the private
EXAMPLE_OUTPUT const and helpers with no visibility edits; the body
carries no depth-sensitive paths (no include_str!/env!).

20 lib tests preserved (identical pre/post); clippy -D warnings and
fmt --check clean; external pub surface byte-identical.

The file crossed 500 lines in PR #306 (prompt optimisation added
tests); the per-session inline-test detector caught it.
